refactor: extract metadata logic out of CLI context (#307)
