+++ tests/tasks/test_metadata.py
@@ -1,0 +1,753 @@
+"""Unit tests for metadata command helper functions."""
+
+from collections.abc import Callable, Iterator
+import json
+from pathlib import Path
+from unittest.mock import AsyncMock, Mock, patch
+
+from lxml import etree  # type: ignore
+import pytest
+
+from docbuild.constants import DEFAULT_DELIVERABLES
+from docbuild.models.deliverable import Deliverable
+from docbuild.models.doctype import Doctype
+from docbuild.models.repo import Repo
+from docbuild.tasks import metadata as metaprocess_pkg
+from docbuild.tasks.metadata import (
+    collect_files_flat,
+    generate_metadata as process,
+    get_deliverable_from_doctype,
+    process_deliverable,
+    process_doctype,
+    store_productdocset_json,
+)
+
+
+@pytest.fixture
+def mock_envconfig(tmp_path: Path) -> Mock:
+    """Provide a mock EnvConfig object with necessary paths and build config."""
+    mock_paths = Mock()
+    mock_paths.repo_dir = tmp_path / "repos"
+    mock_paths.base_cache_dir = tmp_path / "cache"
+    mock_paths.meta_cache_dir = tmp_path / "cache" / "metadata"
+
+    mock_paths.meta_cache_dir.mkdir(parents=True, exist_ok=True)
+
+    mock_paths.tmp_repo_dir = tmp_path / "tmp_repos"
+    mock_paths.main_portal_config = tmp_path / "config" / "dummy.xml"
+
+    mock_tmp = Mock()
+    mock_tmp.tmp_metadata_dir = tmp_path / "tmp" / "metadata"
+    mock_paths.tmp = mock_tmp
+
+    mock_build = Mock()
+    mock_build.daps.meta = "daps-command-template"
+
+    mock_envconfig = Mock()
+    mock_envconfig.paths = mock_paths
+    mock_envconfig.build = mock_build
+    return mock_envconfig
+
+
+@pytest.fixture
+def xmlconfig(request) -> etree.ElementTree:
+    """Parse an XML string into an ElementTree."""
+    xml_string = None
+    if hasattr(request, "param"):
+        xml_string = request.param
+
+    if not xml_string:
+        xml_string = "<docservconfig/>"
+    root = etree.fromstring(xml_string)
+    return etree.ElementTree(root)
+
+
+@pytest.mark.parametrize(
+    "xmlconfig, doctype_str, expected_count, expected_ids",
+    [
+        (
+            """
+            <portal>
+              <product id="sles">
+                <docset id="sles.16-sp6" path="15-sp6">
+                  <resources>
+                    <locale lang="en-us">
+                        <deliverable id="sles.16-sp6.admin">
+                            <dc file="DC-SLE-Micro-5.5-admin">
+                                <format html="1"/>
+                            </dc>
+                        </deliverable>
+                    </locale>
+                  </resources>
+                </docset>
+              </product>
+              <product id="other">
+                <docset id="other.1.0" path="1.0">
+                   <resources>
+                     <locale lang="en-us">
+                        <deliverable>
+                            <dc file="DC-Micro-5.4-cockpit">
+                                <format html="1"/>
+                            </dc>
+                        </deliverable>
+                        <deliverable>
+                            <dc file="DC-Micro-5.5-cockpit">
+                                <format html="1"/>
+                            </dc>
+                        </deliverable>
+                    </locale>
+                   </resources>
+                </docset>
+              </product>
+            </portal>
+            """,
+            "sles/15-sp6/en-us",
+            1,
+            {"sles/15-sp6/en-us:DC-SLE-Micro-5.5-admin"},
+        ),
+        (
+            """
+            <portal>
+              <product id="sles">
+                <docset id="sles.16-sp6" path="15-sp6">
+                    <resources>
+                        <locale lang="en-us">
+                            <deliverable>
+                                <dc file="DC-SLE-Micro-5.5-admin">
+                                    <format html="1"/>
+                                </dc>
+                            </deliverable>
+                        </locale>
+                    </resources>
+                </docset>
+              </product>
+              <product id="other">
+              <docset id="other.1.0" path="1.0">
+                  <resources>
+                    <locale lang="en-us">
+                      <deliverable>
+                        <dc file="DC-Micro-5.4-cockpit">
+                            <format html="1"/>
+                        </dc>
+                      </deliverable>
+                    </locale>
+                  </resources>
+                </docset>
+              </product>
+            </portal>
+            """,
+            "//en-us",
+            2,
+            {
+                "other/1.0/en-us:DC-Micro-5.4-cockpit",
+                "sles/15-sp6/en-us:DC-SLE-Micro-5.5-admin",
+            },
+        ),
+        ("<portal/>", "nonexistent/1.0/en-us", 0, set()),
+        (
+            """<portal>
+                 <product id='sles'>
+                    <docset id='sles.15-sp6' path="15-sp6" />
+                 </product>
+               </portal>""",
+            "sles/15-sp6/de-de",
+            0,
+            set(),
+        ),
+    ],
+    indirect=["xmlconfig"],
+    ids=[
+        "specific_doctype",
+        "wildcard_doctype",
+        "nonexistent_product",
+        "nonexistent_lang",
+    ],
+)
+def test_get_deliverable_from_doctype(
+    xmlconfig, doctype_str, expected_count, expected_ids
+):
+    """Verify deliverables are correctly extracted for various doctypes."""
+    if "nonexistent" in doctype_str:
+        with pytest.raises(ValueError):
+            Doctype.from_str(doctype_str)
+        return
+    else:
+        doctype = Doctype.from_str(doctype_str)
+
+    deliverables = get_deliverable_from_doctype(xmlconfig, doctype)
+
+    assert len(deliverables) == expected_count
+    if expected_ids:
+        assert {d.docsuite for d in deliverables} == expected_ids
+
+
+@pytest.fixture
+def deliverable() -> Deliverable:
+    """Provide a mock Deliverable object for testing."""
+    xml_string = """
+    <docservconfig>
+      <product id="sles">
+        <docset id="sles.15-sp7" path="15-SP7">
+          <resources>
+             <git remote="https://github.com/SUSE/doc-sle.git"/>
+             <locale lang="en-us">
+                <branch>main</branch>
+                <subdir>l10n/sles/en-us</subdir>
+                <deliverable>
+                    <dc file="DC-SLES-deployment">
+                        <format html="1" pdf="1" single-html="0"/>
+                    </dc>
+                </deliverable>
+             </locale>
+          </resources>
+        </docset>
+      </product>
+    </docservconfig>
+    """
+    root = etree.fromstring(xml_string)
+    locale_node = root.find(".//locale")
+    deliverable_node = locale_node.find("deliverable")
+    return Deliverable(deliverable_node)
+
+
+@pytest.fixture
+def stitchnode(deliverable: Deliverable) -> etree._ElementTree:
+    """Build a minimal stitched `docservconfig` ElementTree."""
+    prod_node = etree.Element("product", id=deliverable.xml.productid, productid=deliverable.xml.productid)
+    name_el = etree.SubElement(prod_node, "name")
+    name_el.text = "SUSE Linux Enterprise Server"
+    etree.SubElement(prod_node, "acronym").text = "SLES"
+    etree.SubElement(prod_node, "docset", id=deliverable.xml.docsetid, path=deliverable.xml.docsetid, setid=deliverable.xml.docsetid, productid=deliverable.xml.productid)
+    root = etree.Element("docservconfig")
+    root.append(prod_node)
+    return etree.ElementTree(root)
+
+
+@pytest.mark.asyncio
+class TestProcessDeliverable:
+    """Tests for the process_deliverable function."""
+
+    @pytest.fixture
+    def mock_subprocess(self) -> Iterator[AsyncMock]:
+        """Fixture to mock asyncio.create_subprocess_exec."""
+        with patch.object(
+            metaprocess_pkg.asyncio,
+            "create_subprocess_exec",
+            new_callable=AsyncMock,
+        ) as mock:
+            yield mock
+
+    @pytest.fixture
+    def setup_paths(self, tmp_path: Path, deliverable: Deliverable) -> dict:
+        """Set up common paths and directories for tests."""
+        paths = {
+            "repo_dir": tmp_path / "repos",
+            "tmp_repo_dir": tmp_path / "tmp_repos",
+            "base_cache_dir": tmp_path / "cache",
+            "meta_cache_dir": tmp_path / "cache" / "metadata",
+        }
+        for path in paths.values():
+            path.mkdir(parents=True, exist_ok=True)
+        return paths
+
+    @pytest.mark.parametrize(
+        "scenario, setup_mocks, expected_result, expected_log",
+        [
+            (
+                "success",
+                lambda mocks: None,
+                True,
+                None,
+            ),
+            (
+                "bare_repo_not_found",
+                lambda mocks: mocks["managed_git_repo"].side_effect,
+                False,
+                "Bare repository not found",
+            ),
+            (
+                "clone_fails",
+                lambda mocks: setattr(
+                    mocks["repo_instance"], "clone_bare", AsyncMock(return_value=False)
+                ),
+                False,
+                "Failed to ensure bare repository",
+            ),
+            (
+                "daps_fails",
+                lambda mocks: mocks["subprocess"].__setattr__(
+                    "return_value",
+                    AsyncMock(
+                        returncode=1,
+                        communicate=AsyncMock(return_value=(b"", b"DAPS Error")),
+                    ),
+                ),
+                False,
+                "Error processing",
+            ),
+        ],
+        ids=["success", "bare_repo_not_found", "clone_fails", "daps_fails"],
+    )
+    @patch.object(metaprocess_pkg, "edit_json")
+    @patch.object(metaprocess_pkg, "ManagedGitRepo")
+    async def test_process_deliverable_scenarios(
+        self,
+        mock_managed_git_repo: Mock,
+        mock_edit_json: Mock,
+        deliverable: Deliverable,
+        setup_paths: dict,
+        mock_subprocess: AsyncMock,
+        caplog,
+        scenario: str,
+        setup_mocks: Callable,
+        expected_result: bool,
+        expected_log: str | None,
+    ):
+        """Test various scenarios for process_deliverable."""
+        mock_repo_instance = AsyncMock()
+        mock_managed_git_repo.return_value = mock_repo_instance
+        mock_repo_instance.clone_bare.return_value = True
+        mock_repo_instance.create_worktree.return_value = None
+
+        mock_daps_proc = AsyncMock()
+        mock_daps_proc.communicate.return_value = (b"", b"")
+        mock_daps_proc.returncode = 0
+        mock_subprocess.return_value = mock_daps_proc
+
+        mock_json_data = {"docs": [{"format": {}, "lang": ""}]}
+        mock_edit_json.return_value.__aenter__.return_value = mock_json_data
+        mock_edit_json.return_value.__enter__.return_value = mock_json_data
+
+        mocks = {
+            "managed_git_repo": mock_managed_git_repo,
+            "repo_instance": mock_repo_instance,
+            "subprocess": mock_subprocess,
+        }
+        setup_mocks(mocks)
+
+        if scenario != "bare_repo_not_found":
+            (setup_paths["repo_dir"] / deliverable.git.slug).mkdir()
+
+        dapstmpl = "daps --dc-file={dcfile} --output={output}"
+
+        result = await process_deliverable(
+            deliverable=deliverable,
+            repo_dir=setup_paths["repo_dir"],
+            tmp_repo_dir=setup_paths["tmp_repo_dir"],
+            meta_cache_dir=setup_paths["meta_cache_dir"],
+            dapstmpl=dapstmpl,
+        )
+
+        success, res_deliverable = result
+        assert success is expected_result
+        assert res_deliverable == deliverable
+
+        if expected_log:
+            assert any(expected_log in record.message for record in caplog.records)
+
+
+@pytest.mark.asyncio
+class TestProcessDoctype:
+    """Tests for the process_doctype function."""
+
+    @pytest.fixture
+    def mock_root(self) -> etree._ElementTree:
+        """Provide a mock XML root element for testing."""
+        return etree.ElementTree(etree.fromstring("<docservconfig/>"))
+
+    @patch.object(metaprocess_pkg, "process_deliverable", new_callable=AsyncMock)
+    @patch.object(metaprocess_pkg, "get_deliverable_from_doctype")
+    async def test_success_with_deliverables(
+        self,
+        mock_get_deliverables: Mock,
+        mock_process_deliverable: AsyncMock,
+        mock_root: etree._ElementTree,
+        mock_envconfig: Mock,
+    ):
+        """Test successful processing when deliverables are found."""
+        doctype = Doctype.from_str("sles/15/en-us")
+        mock_deliverable = Mock(
+            spec=Deliverable, git=Mock(spec=Repo, url="gh://SUSE/doc-test")
+        )
+        mock_deliverables = [mock_deliverable, mock_deliverable]
+        mock_get_deliverables.return_value = mock_deliverables
+        mock_process_deliverable.return_value = True
+
+        env = mock_envconfig.paths
+        result = await process_doctype(
+            mock_root,
+            doctype,
+            repo_dir=env.repo_dir,
+            tmp_repo_dir=env.tmp_repo_dir,
+            meta_cache_dir=env.meta_cache_dir,
+            dapsmetatmpl=mock_envconfig.build.daps.meta,
+            exitfirst=False
+        )
+
+        assert not result
+        mock_get_deliverables.assert_called_once_with(mock_root, doctype)
+        assert mock_process_deliverable.call_count == 2
+
+    @patch.object(metaprocess_pkg, "process_deliverable", new_callable=AsyncMock)
+    @patch.object(metaprocess_pkg, "get_deliverable_from_doctype")
+    async def test_no_deliverables_found(
+        self,
+        mock_get_deliverables: Mock,
+        mock_process_deliverable: AsyncMock,
+        mock_root: etree._ElementTree,
+        mock_envconfig: Mock,
+    ):
+        """Test behavior when no deliverables are found for the doctype."""
+        doctype = Doctype.from_str("sles/15/en-us")
+        mock_get_deliverables.return_value = []
+
+        env = mock_envconfig.paths
+        result = await process_doctype(
+            mock_root, doctype, env.repo_dir, env.tmp_repo_dir, env.meta_cache_dir, mock_envconfig.build.daps.meta
+        )
+
+        assert not result
+        mock_get_deliverables.assert_called_once_with(mock_root, doctype)
+        mock_process_deliverable.assert_not_called()
+
+    @patch.object(metaprocess_pkg, "process_deliverable", new_callable=AsyncMock)
+    @patch.object(metaprocess_pkg, "get_deliverable_from_doctype")
+    @patch.object(metaprocess_pkg, "update_repositories", new_callable=AsyncMock)
+    async def test_exitfirst_fails_fast(
+        self,
+        mock_update_repositories: AsyncMock,
+        mock_get_deliverables: Mock,
+        mock_process_deliverable: AsyncMock,
+        mock_root: etree._ElementTree,
+        mock_envconfig: Mock,
+    ):
+        """When `exitfirst=True`, process_doctype should stop on first failure."""
+        doctype = Doctype.from_str("sles/15/en-us")
+
+        d1 = Mock(spec=Deliverable)
+        d1.full_id = "sles/15/en-us:DC-ONE"
+        d2 = Mock(spec=Deliverable)
+        d2.full_id = "sles/15-en-us:DC-TWO"
+
+        mock_get_deliverables.return_value = [d1, d2]
+        mock_process_deliverable.side_effect = [(False, d1), (True, d2)]
+        mock_update_repositories.return_value = None
+
+        env = mock_envconfig.paths
+        failed = await process_doctype(
+            mock_root, doctype, env.repo_dir, env.tmp_repo_dir, env.meta_cache_dir, mock_envconfig.build.daps.meta, exitfirst=True
+        )
+
+        assert failed == [d1]
+
+
+@pytest.mark.asyncio
+class TestProcessEmptyDoctypes:
+    """Tests for the case when no doctypes are passed to process."""
+
+    @patch.object(metaprocess_pkg, "store_productdocset_json", new_callable=Mock)
+    @patch.object(metaprocess_pkg, "collect_files_flat", new_callable=Mock)
+    @patch.object(metaprocess_pkg, "parse_portal_config", new_callable=AsyncMock)
+    @patch.object(metaprocess_pkg, "process_doctype", new_callable=AsyncMock)
+    async def test_process_empty_doctypes(
+        self,
+        mock_process_doctype: AsyncMock,
+        mock_parse_portal_config: AsyncMock,
+        mock_collect_files_flat: Mock,
+        mock_store_json: Mock,
+        mock_envconfig: Mock,
+    ):
+        """Test process function with an empty tuple of doctypes."""
+        xml_string = """
+        <docservconfig>
+            <product id="sles">
+              <name>SUSE Linux Enterprise Server</name>
+              <acronym>SLES</acronym>
+              <docset id="sles.15-sp6" path="15-SP6"/>
+            </product>
+        </docservconfig>
+        """
+        mock_stitch_node = etree.ElementTree(etree.fromstring(xml_string))
+        mock_parse_portal_config.return_value = mock_stitch_node
+
+        mock_collect_files_flat.return_value = [
+            (Doctype.from_str(DEFAULT_DELIVERABLES), "*", [Path("dummy.xml")])
+        ]
+
+        env = mock_envconfig.paths
+        await process(
+            env.main_portal_config, env.tmp.tmp_metadata_dir, env.repo_dir,
+            env.tmp_repo_dir, env.meta_cache_dir, mock_envconfig.build.daps.meta, ()
+        )
+
+        mock_parse_portal_config.assert_called_once()
+        mock_process_doctype.assert_called()
+        mock_store_json.assert_called()
+
+    @patch.object(metaprocess_pkg, "store_productdocset_json", new_callable=Mock)
+    @patch.object(metaprocess_pkg, "collect_files_flat", new_callable=Mock)
+    @patch.object(metaprocess_pkg, "parse_portal_config", new_callable=AsyncMock)
+    @patch.object(metaprocess_pkg, "process_doctype", new_callable=AsyncMock)
+    async def test_no_doctypes_uses_default(
+        self,
+        mock_process_doctype: AsyncMock,
+        mock_parse_portal_config: AsyncMock,
+        mock_collect_files_flat: Mock,
+        mock_store_json: Mock,
+        mock_envconfig: Mock,
+    ):
+        """Test process uses a default doctype when none are provided."""
+        xml_string = """
+        <docservconfig>
+            <product id="sles">
+              <name>SUSE Linux Enterprise Server</name>
+              <acronym>SLES</acronym>
+              <docset id="sles.15-sp6" path="15-SP6"/>
+            </product>
+        </docservconfig>
+        """
+        mock_stitch_node = etree.ElementTree(etree.fromstring(xml_string))
+
+        mock_parse_portal_config.return_value = mock_stitch_node
+        mock_process_doctype.return_value = []
+        mock_collect_files_flat.return_value = [
+            (Doctype.from_str(DEFAULT_DELIVERABLES), "*", [Path("dummy.xml")])
+        ]
+
+        env = mock_envconfig.paths
+        with patch.object(metaprocess_pkg, "stdout"):
+            result = await process(
+                env.main_portal_config, env.tmp.tmp_metadata_dir, env.repo_dir,
+                env.tmp_repo_dir, env.meta_cache_dir, mock_envconfig.build.daps.meta, ()
+            )
+
+        assert result == 0
+        mock_parse_portal_config.assert_awaited_once()
+        mock_store_json.assert_called()
+        default_doctype = Doctype.from_str(DEFAULT_DELIVERABLES)
+        mock_process_doctype.assert_awaited_once_with(
+            mock_stitch_node,
+            default_doctype,
+            env.repo_dir,
+            env.tmp_repo_dir,
+            env.meta_cache_dir,
+            mock_envconfig.build.daps.meta,
+            exitfirst=False,
+            skip_repo_update=False,
+        )
+
+    @patch.object(metaprocess_pkg, "store_productdocset_json", new_callable=Mock)
+    @patch.object(metaprocess_pkg, "collect_files_flat", new_callable=Mock)
+    @patch.object(metaprocess_pkg, "parse_portal_config", new_callable=AsyncMock)
+    @patch.object(metaprocess_pkg, "process_doctype", new_callable=AsyncMock)
+    async def test_process_reports_failed_deliverables_and_returns_one(
+        self,
+        mock_process_doctype: AsyncMock,
+        mock_parse_portal_config: AsyncMock,
+        mock_collect_files_flat: Mock,
+        mock_store_json: Mock,
+        mock_envconfig: Mock,
+        deliverable: Deliverable,
+        stitchnode: etree._ElementTree,
+    ):
+        """When any deliverable fails, `process` should report and return 1."""
+        mock_parse_portal_config.return_value = stitchnode
+
+        mock_deliverable = Mock(spec=Deliverable)
+        mock_deliverable.full_id = (
+            f"{deliverable.xml.productid}/{deliverable.xml.docsetid}/{deliverable.xml.lang}:DC-FAIL"
+        )
+        mock_process_doctype.return_value = [mock_deliverable]
+        mock_collect_files_flat.return_value = []
+
+        env = mock_envconfig.paths
+        with patch.object(metaprocess_pkg, "console_err") as mock_console_err:
+            result = await process(
+                env.main_portal_config, env.tmp.tmp_metadata_dir, env.repo_dir,
+                env.tmp_repo_dir, env.meta_cache_dir, mock_envconfig.build.daps.meta, ()
+            )
+
+        assert result == 1
+        assert mock_console_err.print.called
+
+
+def test_store_productdocset_json_merges_and_writes(
+    mock_envconfig: Mock,
+    deliverable: Deliverable,
+    stitchnode: etree._ElementTree,
+):
+    """Merge docs from metadata files and write product/docset JSON."""
+    meta_cache_dir = mock_envconfig.paths.meta_cache_dir
+    meta_file = Path(meta_cache_dir) / "meta1.json"
+    doc_content = {
+        "docs": [
+            {
+                "title": "Doc1",
+                "dcfile": "DC-Doc1.xml",
+                "lang": "en-us",
+                "description": "A test document.",
+                "dateModified": "2024-01-01",
+                "format": {"html": "path/to/html"},
+            }
+        ]
+    }
+    meta_file.write_text(json.dumps(doc_content), encoding="utf-8")
+
+    doctype = Doctype.from_str(
+        f"{deliverable.xml.productid}/{deliverable.xml.docsetid}/{deliverable.xml.lang}"
+    )
+
+    with patch.object(
+        metaprocess_pkg,
+        "collect_files_flat",
+        return_value=[(doctype, deliverable.xml.docsetid, [Path("meta1.json")])],
+    ):
+        store_productdocset_json([doctype], stitchnode, meta_cache_dir)
+
+    out_file = Path(meta_cache_dir) / deliverable.xml.productid / f"{deliverable.xml.docsetid}.json"
+    assert out_file.exists()
+    merged = json.loads(out_file.read_text(encoding="utf-8"))
+    assert "documents" in merged
+    assert merged["documents"][0]["docs"][0]["title"] == "Doc1"
+    assert "hide-productname" in merged
+
+
+def test_store_productdocset_json_warns_on_empty_metadata(
+    mock_envconfig: Mock,
+    deliverable: Deliverable,
+    stitchnode: etree._ElementTree,
+):
+    """If a metadata file is empty ({}), a warning is logged."""
+    meta_cache_dir = mock_envconfig.paths.meta_cache_dir
+    meta_file = meta_cache_dir / "empty.json"
+    meta_file.write_text("{}", encoding="utf-8")
+
+    doctype = Doctype.from_str(
+        f"{deliverable.xml.productid}/{deliverable.xml.docsetid}/{deliverable.xml.lang}"
+    )
+
+    with (
+        patch.object(
+            metaprocess_pkg,
+            "collect_files_flat",
+            return_value=[(doctype, deliverable.xml.docsetid, [Path("empty.json")])],
+        ),
+        patch.object(metaprocess_pkg, "log") as mock_log,
+    ):
+        store_productdocset_json([doctype], stitchnode, meta_cache_dir)
+
+    mock_log.error.assert_called_with("Empty metadata file %s", Path("empty.json"))
+
+
+def test_store_productdocset_json_handles_read_error(
+    mock_envconfig: Mock,
+    deliverable: Deliverable,
+    stitchnode: etree._ElementTree,
+):
+    """If reading a metadata file raises, it should be caught and an error logged."""
+    meta_cache_dir = mock_envconfig.paths.meta_cache_dir
+    bad_file = meta_cache_dir / "bad.json"
+    bad_file.write_text("{ not json }", encoding="utf-8")
+
+    doctype = Doctype.from_str(
+        f"{deliverable.xml.productid}/{deliverable.xml.docsetid}/{deliverable.xml.lang}"
+    )
+
+    with (
+        patch.object(
+            metaprocess_pkg,
+            "collect_files_flat",
+            return_value=[(doctype, deliverable.xml.docsetid, [Path("bad.json")])],
+        ),
+        patch.object(metaprocess_pkg, "log") as mock_log,
+    ):
+        store_productdocset_json([doctype], stitchnode, meta_cache_dir)
+
+    mock_log.error.assert_called()
+
+
+@pytest.mark.parametrize(
+    "setup_files, doctype_str, expected_file_count",
+    [
+        (
+            {"en-us/sles/15-SP4": ["DC-file1", "DC-file2", "ignored.xml"]},
+            "sles/15-SP4/en-us",
+            2,
+        ),
+        (
+            {
+                "en-us/sles/15-SP4": ["DC-file-foo", "DC-file-bar"],
+                "de-de/sles/15-SP4": ["DC-file-foo", "DC-file-bar"],
+            },
+            "sles/15-SP4/en-us,de-de",
+            4,
+        ),
+        (
+            {},
+            "sles/15-SP4/en-us",
+            0,
+        ),
+    ],
+    ids=["single_lang", "multi_lang", "no_files"],
+)
+def test_collect_files_flat(
+    tmp_path: Path, setup_files, doctype_str, expected_file_count
+):
+    """Verify that collect_files_flat finds DC-* files correctly."""
+    cache_dir = tmp_path / "cache"
+    for path_str, files in setup_files.items():
+        dir_path = cache_dir / path_str
+        dir_path.mkdir(parents=True, exist_ok=True)
+        for f in files:
+            (dir_path / f).touch()
+
+    doctypes = [Doctype.from_str(doctype_str)]
+
+    results = list(collect_files_flat(doctypes, cache_dir))
+
+    assert len(results) == (1 if expected_file_count > 0 else 0)
+    if results:
+        _, _, found_files = results[0]
+        assert len(found_files) == expected_file_count
+
+
+@pytest.mark.asyncio
+class TestUpdateRepositories:
+    """Tests for the update_repositories function."""
+
+    @patch.object(metaprocess_pkg.ManagedGitRepo, "clone_bare", new_callable=AsyncMock)
+    async def test_update_repositories_success(
+        self, mock_clone_bare: AsyncMock, tmp_path: Path
+    ):
+        """Verify that update_repositories successfully 'clones' a repo."""
+        mock_deliverable = Mock(spec=Deliverable)
+        mock_deliverable.git.url = "gh://SUSE/doc-test"
+        mock_deliverable.git.slug = "SUSE-doc-test"
+        deliverables = [mock_deliverable]
+        repo_dir = tmp_path / "repos"
+        mock_clone_bare.return_value = True
+
+        from docbuild.tasks.metadata import update_repositories
+
+        await update_repositories(deliverables, repo_dir) # type: ignore[arg-type]
+
+        mock_clone_bare.assert_awaited_once()
+
+    @patch.object(metaprocess_pkg.ManagedGitRepo, "clone_bare", new_callable=AsyncMock)
+    async def test_update_repositories_failed(
+        self, mock_clone_bare: AsyncMock, tmp_path: Path, caplog
+    ):
+        """Verify that update_repositories handles a git clone failure."""
+        mock_deliverable = Mock(spec=Deliverable)
+        mock_deliverable.git.url = "gh://SUSE/non-existent-repo"
+        mock_deliverable.git.slug = "SUSE-non-existent-repo"
+        deliverables = [mock_deliverable]
+        repo_dir = tmp_path / "repos"
+        mock_clone_bare.return_value = False
+
+        from docbuild.tasks.metadata import update_repositories
+
+        await update_repositories(deliverables, repo_dir) # type: ignore[arg-type]
+
+        mock_clone_bare.assert_awaited_once()
+        assert "Failed to update" in caplog.text
